+++ src/widgets/antd/map.jsx
@@ -1,2 +1,2 @@
-import map from '../common/map';
+import map from '../../components/map';
 export default map;
